% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoplot.scc_chart.R
\name{autoplot.scc_chart}
\alias{autoplot.scc_chart}
\title{Plot a SCC chart with \pkg{ggplot2}.}
\usage{
autoplot.scc_chart(object, xlab = "block", ylab = object$y_value_name,
  title = object$description, append.hist = FALSE, ...)
}
\arguments{
\item{object}{[\code{scc_chart}]\cr
Chart for plotitng.}

\item{xlab}{[\code{character(1)}]\cr
A title for the x axis.}

\item{ylab}{[\code{character(1)}]\cr
A title for the y axis.}

\item{title}{[\code{character(1)}]\cr
A title for the plot. See main parameter in plot function.}

\item{append.hist}{[\code{logical(1)}]\cr
Should a histogram of the y-value be depicted beside the chart?
Default ist \code{FALSE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
Takes a \code{scc_chart} object and generates a graphical representation
of the chart in form of a \code{\link[ggplot2]{ggplot}} object.
}

