\name{generate_random_blocks}
\alias{generate_random_blocks}
\title{Function for generating random blocks}
\usage{
generate_random_blocks(n = 10, block_size = 5)
}
\arguments{
  \item{n}{[integer(1)]\cr Number of blocks.}

  \item{block_size}{[integer]\cr Size of each block.}
}
\description{
Function for generating random blocks
}

