% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateRandomBlocks.R
\name{generateRandomBlocks}
\alias{generateRandomBlocks}
\title{Function for generating random blocks}
\usage{
generateRandomBlocks(n = 10, block.size = 5)
}
\arguments{
\item{n}{[integer(1)]\cr
Number of blocks.}

\item{block.size}{[integer]\cr
Size of each block.}
}
\value{
List of blocks.
}
\description{
Function for generating random blocks
}

