% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getLongestOutlierSequence.R
\name{getLongestOutlierSequence}
\alias{getLongestOutlierSequence}
\title{Determines the length of the longest consecutive outlier sequence.}
\usage{
getLongestOutlierSequence(x, ...)
}
\arguments{
\item{x}{[\code{scc_chart}]\cr
Chart.}

\item{...}{[\code{list}]\cr
Not used.}
}
\value{
List containing the length and the starting index.
 Length is -1, if the are no observations violating the
 control bounds.
}
\description{
Determines the length of the longest consecutive outlier sequence.
}

