% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNumberOfBlocks.R
\name{getNumberOfBlocks}
\alias{getNumberOfBlocks}
\title{Returns number of blocks of a chart.}
\usage{
getNumberOfBlocks(x)
}
\arguments{
\item{x}{[\code{scc_chart}]\cr
  An object of type \code{scc_chart}.}
}
\value{
[\code{integer(1)}]
   Number of blocks.
}
\description{
Returns number of blocks of a chart.
}

