% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNumberOfOutliers.R
\name{getNumberOfOutliers}
\alias{getNumberOfOutliers}
\title{Determines the number of outliers, i.e., values out of town.}
\usage{
getNumberOfOutliers(x, ...)
}
\arguments{
\item{x}{[\code{scc_chart}]\cr
Chart.}

\item{...}{[any]\cr
Not used.}
}
\value{
Data frame containing the number of y-values below LCB and UCB.
}
\description{
Determines the number of outliers, i.e., values out of town.
}

