\name{number_of_blocks}
\alias{number_of_blocks}
\title{Returns number of blocks of a chart.}
\usage{
number_of_blocks(x)
}
\arguments{
  \item{x}{[\code{\link{scc_chart}}]\cr An object of type
  \code{\link{scc_chart}}.}
}
\value{
[\code{integer(1)}]\cr Number of blocks.
}
\description{
Returns number of blocks of a chart.
}

