\name{number_of_outliers}
\alias{number_of_outliers}
\title{Determines the number of outliers, i.e., values out of town.}
\usage{
number_of_outliers(x, ...)
}
\arguments{
  \item{x}{[\code{\link{scc_chart}}]\cr An object of type
  \code{\link{scc_chart}}.}

  \item{...\cr}{Not used.}
}
\value{
Data frame containing the number of y-values below LCB and
UCB.
}
\description{
Determines the number of outliers, i.e., values out of
town.
}

