\name{print.scc_chart}
\alias{print.scc_chart}
\title{Print a SCC chart.}
\usage{
\method{print}{scc_chart}(x, ...)
}
\arguments{
  \item{x}{[\code{\link{scc_chart}}]\cr A
  \code{\link{scc_chart}} object.}

  \item{...\cr}{Not used.}
}
\value{
Nothing. As a side-effect a textual information is printed
to the console.
}
\description{
Print a SCC chart.
}

