\name{number_of_blocks}
\alias{number_of_blocks}
\title{Returns number of blocks of a chart.}
\usage{
number_of_blocks(x)
}
\arguments{
  \item{x}{[\code{scc_chart}]\cr An object of type
  \code{scc_chart}.}
}
\value{
[\code{integer(1)}] Number of blocks.
}
\description{
Returns number of blocks of a chart.
}

