% Generated by roxygen2 (4.0.1): do not edit by hand
\name{number_of_outliers}
\alias{number_of_outliers}
\title{Determines the number of outliers, i.e., values out of town.}
\usage{
number_of_outliers(x, ...)
}
\arguments{
\item{x}{[\code{scc_chart}]\cr
An object of type \code{scc_chart}.}

\item{...}{[\code{list}]\cr
  Not used.}
}
\value{
Data frame containing the number of y-values below LCB and UCB.
}
\description{
Determines the number of outliers, i.e., values out of town.
}

