% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.scc_chart}
\alias{plot.scc_chart}
\title{Plot a SCC chart.}
\usage{
\method{plot}{scc_chart}(x, ...)
}
\arguments{
\item{x}{[\code{scc_chart}]\cr
An object of type \code{scc_chart}.}

\item{...}{[\code{list}]\cr
  Further params passed to plot function.}
}
\value{
Nothing. As a side-effect the chart is plotted.
}
\description{
Plot a SCC chart.
}

