% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.scc_chart.R
\name{print.scc_chart}
\alias{print.scc_chart}
\title{Print a SCC chart.}
\usage{
\method{print}{scc_chart}(x, ...)
}
\arguments{
\item{x}{[\code{scc_chart}]\cr
A chart.}

\item{...}{[any]\cr
Not used.}
}
\value{
Nothing. As a side-effect a textual information is printed to the console.
}
\description{
Print a SCC chart.
}

